決算日セクションを /market から /disclosure に移設 + タブ名「決算・開示」へ (#213)
- 決算日カード (編集UI込み) と /api/kessan_comment GET/POST を disclosure_bp に移植
- /market から決算日セクションを完全に取り除き、市場→テーマランク→テーマニュースの並びだけにする
- ナビタブ「適宜開示」→「決算・開示」
- helpers.get_market_html_parts: プレースホルダ挿入を撤去し戻り値キーを body にリネーム (決算日 h2 抑制は継続)
- テスト: /market 経由の決算カード assert を /disclosure 経由に振り替え、ナビ assert をタブ名追随
