+++ scripts/webapp/templates/base.html
@@ -15,7 +15,7 @@
     <li><a href="/" class="brand">Shintakane Research</a></li>
     <li><a href="/market" class="nav-link">市場データ</a></li>
     <li><a href="/portfolio" class="nav-link">保有銘柄</a></li>
-    <li><a href="/disclosure" class="nav-link">適宜開示</a></li>
+    <li><a href="/disclosure" class="nav-link">決算・開示</a></li>
   </ul>
   <ul>
     <li class="quick-search">
