+++ scripts/webapp/templates/disclosure.html
@@ -1,5 +1,6 @@
 {% extends "base.html" %}
-{% block title %}適宜開示 - Shintakane Research{% endblock %}
+{% import "_macros.html" as macros %}
+{% block title %}決算・開示 - Shintakane Research{% endblock %}
 {% block content %}
 
 {% if disclosure_parts.available %}
@@ -8,13 +9,155 @@
 {{ disclosure_parts.css | safe }}
 </style>
 {{ disclosure_parts.header | safe }}
-{{ disclosure_parts.body | safe }}
-{{ disclosure_parts.footer | safe }}
 {% else %}
 <div style="padding:12px;margin:12px 0;background:#fef3e6;border:1px solid #f5c37c;border-radius:6px;">
-  適宜開示データ (<code>disclosure_data.html</code>) が未生成です。<br>
-  <small>scripts/make_market_db.py を実行すると生成されます。</small>
+  決算・開示データ (<code>disclosure_data.html</code>) が未生成です。<br>
+  <small>scripts/make_market_db.py を実行すると生成されます。決算カレンダーは以下の通り利用できます。</small>
 </div>
+{% endif %}
+
+{# 動的決算セクション (市場ページから移設) #}
+<h2>決算日</h2>
+<div id="kessan-section">
+  {% macro render_kessan_editor(s, is_past) -%}
+    <div class="kessan-editor" style="display:none;" data-loaded="0">
+      <div class="kessan-editor-row">
+        <label>期待度:</label>
+        <select class="kessan-field kessan-pre-expectation">
+          {% for opt in expectation_options %}
+            <option value="{{ opt }}">{{ opt or '—' }}</option>
+          {% endfor %}
+        </select>
+      </div>
+      <div class="kessan-editor-row">
+        <label>事前見通し:</label>
+        <textarea class="kessan-field kessan-pre-outlook" rows="2" placeholder="決算前の見通し"></textarea>
+      </div>
+      {% if is_past %}
+      <div class="kessan-editor-row">
+        <label>株価変動率:</label>
+        <span class="kessan-field-display kessan-post-price-change">
+          <span class="kessan-pc-1d">{{ s.post_price_changes['1d'] or '-' }}</span>
+          /
+          <span class="kessan-pc-5d">{{ s.post_price_changes['5d'] or '-' }}</span>
+        </span>
+      </div>
+      <div class="kessan-editor-row">
+        <label>反応コメ:</label>
+        <textarea class="kessan-field kessan-post-comment" rows="2" placeholder="決算後の反応・コメント"></textarea>
+      </div>
+      <div class="kessan-editor-row">
+        <label>決算またぎ:</label>
+        <input type="checkbox" class="kessan-field kessan-matagi">
+      </div>
+      {% endif %}
+    </div>
+  {%- endmacro %}
+
+  {% macro render_stock_li(s, is_past) -%}
+    {% set li_class = 'kessan-4q' if (s.quarter|string) == '4' else '' %}
+    {% set possess_class = ' is-possess' if s.is_possess else '' %}
+    <li class="{{ li_class }} kessan-stock{% if s.has_comment %} has-comment{% endif %}{{ possess_class }}"
+        data-code="{{ s.code_s }}"
+        data-kessanbi="{{ s.kessanbi }}"
+        data-quarter="{{ s.quarter }}"
+        data-is-past="{{ '1' if is_past else '0' }}"
+        onclick="handleKessanRowClick(event, this)"
+        title="クリックでコメント編集">
+      {% if s.is_possess %}<span class="kessan-possess-mark" title="保有銘柄">☆</span>{% endif %}
+      {% if s.kessan_matagi %}<span class="kessan-matagi-mark" title="決算またぎ保有">◆</span>{% endif %}
+      <a href="https://kabutan.jp/stock/chart?code={{ s.code_s }}" target="_blank">{{ s.code_s }}</a>
+      <a href="/stock/{{ s.code_s }}">{% if s.stock_name %}{{ macros.stock_name_with_prev(s) }}{% else %}-{% endif %}</a>
+      {% if s.quarter %}<span class="kessan-q-label">[{{ s.quarter }}Q]</span>{% endif %}
+      {% if s.pre_expectation %}<span class="kessan-expectation-badge exp-{{ s.pre_expectation }}">{{ s.pre_expectation }}</span>{% endif %}
+      {% if is_past and (s.post_price_changes['pts'] or s.post_price_changes['1d'] or s.post_price_changes['5d']) %}
+        {% set pcp = s.post_price_changes['pts'] %}
+        {% set pc1 = s.post_price_changes['1d'] %}
+        {% set pc5 = s.post_price_changes['5d'] %}
+        <span class="kessan-price-change">
+          {% if pcp %}<span class="kessan-pts-label">PTS</span><span class="{% if pcp.startswith('-') %}rate-neg{% else %}rate-pos{% endif %}">{{ pcp }}%</span>{% endif %}
+          {% if pc1 or pc5 %}
+            {% if pc1 %}<span class="{% if pc1.startswith('-') %}rate-neg{% else %}rate-pos{% endif %}">{{ pc1 }}%</span>{% else %}<span>-</span>{% endif %}
+            /
+            {% if pc5 %}<span class="{% if pc5.startswith('-') %}rate-neg{% else %}rate-pos{% endif %}">{{ pc5 }}%</span>{% else %}<span>-</span>{% endif %}
+          {% endif %}
+        </span>
+      {% endif %}
+      {% if s.pre_outlook or s.post_comment %}
+      <div class="kessan-comment-view">
+        {% if s.pre_outlook %}<div class="kessan-pre-view"><span class="kessan-view-label">見通し:</span> {{ s.pre_outlook }}</div>{% endif %}
+        {% if is_past and s.post_comment %}<div class="kessan-post-view"><span class="kessan-view-label">反応:</span> {{ s.post_comment }}</div>{% endif %}
+      </div>
+      {% endif %}
+      {{ render_kessan_editor(s, is_past) }}
+    </li>
+  {%- endmacro %}
+
+  {# 直近1週間の過去決算: 常時表示 #}
+  {% if kessan_data.recent_past_entries %}
+  <div class="kessan-grid">
+    {% for kessanbi, stocks in kessan_data.recent_past_entries %}
+    <div class="kessan-card past">
+      <div class="card-date">{{ kessanbi[5:] }} (済)</div>
+      <ul class="stock-list">
+        {% for s in stocks %}{{ render_stock_li(s, true) }}{% endfor %}
+      </ul>
+    </div>
+    {% endfor %}
+  </div>
+  {% endif %}
+
+  {# 当日決算 + 未来決算: 同じグリッドに横並び表示。
+     当日カードは中身は past 扱い (is_past=true) で反応コメ・株価変動率・
+     決算またぎを当日中に編集可能。見た目は通常 (past クラスなし、"(済)" ラベルなし)。
+     未来カードは future クラスで事前見通しのみ編集可。 #}
+  {% if kessan_data.today_entries or kessan_data.future_entries %}
+  <div class="kessan-grid">
+    {% for kessanbi, stocks in kessan_data.today_entries %}
+    <div class="kessan-card">
+      <div class="card-date">{{ kessanbi[5:] }}</div>
+      <ul class="stock-list">
+        {% for s in stocks %}{{ render_stock_li(s, true) }}{% endfor %}
+      </ul>
+    </div>
+    {% endfor %}
+    {% for kessanbi, stocks in kessan_data.future_entries %}
+    <div class="kessan-card future">
+      <div class="card-date">{{ kessanbi[5:] }}</div>
+      <ul class="stock-list">
+        {% for s in stocks %}{{ render_stock_li(s, false) }}{% endfor %}
+      </ul>
+    </div>
+    {% endfor %}
+  </div>
+  {% endif %}
+
+  {# それ以前の過去決算: 折りたたみ #}
+  {% if kessan_data.older_past_entries %}
+  <details>
+    <summary style="cursor:pointer; color:#666; margin:8px 0;">▶ さらに前の決算を表示 ({{ kessan_data.older_past_entries|length }}件)</summary>
+    <div class="kessan-grid">
+      {% for kessanbi, stocks in kessan_data.older_past_entries %}
+      <div class="kessan-card past">
+        <div class="card-date">{{ kessanbi[5:] }} (済)</div>
+        <ul class="stock-list">
+          {% for s in stocks %}{{ render_stock_li(s, true) }}{% endfor %}
+        </ul>
+      </div>
+      {% endfor %}
+    </div>
+  </details>
+  {% endif %}
+
+  {% if not kessan_data.future_entries and not kessan_data.past_entries %}
+  <div style="color:#888;padding:12px;">決算予定の銘柄がありません（ポートフォリオ設定を確認してください）。</div>
+  {% endif %}
+</div>
+
+{# 適宜開示 (自動収集) #}
+{% if disclosure_parts.available %}
+{{ disclosure_parts.body | safe }}
+{{ disclosure_parts.footer | safe }}
 {% endif %}
 
 {% endblock %}
